# tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "2009scape Wiki API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/entities/{entity_type}/{ref}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get the whole page for one entity",
        "description": "Get everything needed to draw one entity's page: identity, the infobox, the\nattribute sections, and a first page of each set of related entities.",
        "operationId": "entity",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
            },
            "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
          },
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id.",
              "title": "Ref"
            },
            "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id."
          }
        ],
        "responses": {
          "200": {
            "description": "The page, described as data and ready to render.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDescriptor"
                }
              }
            }
          },
          "404": {
            "description": "Nothing published answers to that reference. The code says which case it is: `not_found` means there is no such entity, `not_published` means it exists in this build but is held back. A `not_found` for a name rather than an id also carries `near_names`, the address that answers what the name might have been meant to be.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "308": {
            "description": "That reference is retired. `Location` holds the URL the same entity answers at now. Follow it, and store the new one."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/entities/{entity_type}/{ref}/tooltip": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get the hover card for one entity",
        "description": "Get identity, one sentence, and the few values the registry marks for hover.",
        "operationId": "tooltip",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
            },
            "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
          },
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id.",
              "title": "Ref"
            },
            "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id."
          }
        ],
        "responses": {
          "200": {
            "description": "Just enough about the entity to fill a hover card.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tooltip"
                }
              }
            }
          },
          "404": {
            "description": "Nothing published answers to that reference. The code says which case it is: `not_found` means there is no such entity, `not_published` means it exists in this build but is held back. A `not_found` for a name rather than an id also carries `near_names`, the address that answers what the name might have been meant to be.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "308": {
            "description": "That reference is retired. `Location` holds the URL the same entity answers at now. Follow it, and store the new one."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
...
    "/v1/entities/{entity_type}/{ref}/resolve": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Check what a reference points at, without being sent there",
        "description": "Ask what a reference points at without being redirected there.\n\nLive, retired, held back and unknown all come back as a 200 with an `outcome`.",
        "operationId": "resolve",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
            },
            "description": "Which type of thing to look up, such as `item` or `npc`. `GET /v1/types` lists the types this build serves."
          },
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id.",
              "title": "Ref"
            },
            "description": "How to address the entity: its id (`4587`), its slug (`dragon-scimitar`), a slug it used to have, or the token the game source calls it by. A reference made only of digits is always an id."
          }
        ],
        "responses": {
          "200": {
            "description": "What the reference resolves to, stated in the body.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Present"
                    },
                    {
                      "$ref": "#/components/schemas/Moved"
                    },
                    {
                      "$ref": "#/components/schemas/Hidden"
                    },
                    {
                      "$ref": "#/components/schemas/Missing"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "outcome",
                    "mapping": {
                      "found": "#/components/schemas/Present",
                      "moved": "#/components/schemas/Moved",
                      "hidden": "#/components/schemas/Hidden",
                      "missing": "#/components/schemas/Missing"
                    }
                  },
                  "title": "Response Resolve"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/types": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List every entity type and how its fields are presented",
        "description": "List every type with its attributes and its relationships, labelled for reading\neach way round.",
        "operationId": "types",
        "responses": {
          "200": {
            "description": "Every type served, with its attributes and relationships.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TypeInfo"
                  },
                  "type": "array",
                  "title": "Response Types"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
...
    "/v1/near-names": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Ask which real names a misspelt one might have meant",
        "description": "Turn an unmatched name into the real names it may have meant, each carrying\nidentity and nothing else.\n\nAn empty answer means nothing was close, which is an answer rather than a failure.",
        "operationId": "near_names",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The name that answered to nothing, exactly as it was typed, such as `dragon scimtar`.",
              "title": "Name"
            },
            "description": "The name that answered to nothing, exactly as it was typed, such as `dragon scimtar`."
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which sort of thing the name belongs to. Required: `dagon scimitar` and `dagon` are near misses for different things depending on whether an item or an npc was meant, so this question cannot be answered without it. `GET /v1/types` lists the sorts this build serves."
            },
            "description": "Which sort of thing the name belongs to. Required: `dagon scimitar` and `dagon` are near misses for different things depending on whether an item or an npc was meant, so this question cannot be answered without it. `GET /v1/types` lists the sorts this build serves."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 25,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "At most how many close names to offer. Deliberately small: this is a question to put to a person, not a listing to read.",
              "title": "Limit"
            },
            "description": "At most how many close names to offer. Deliberately small: this is a question to put to a person, not a listing to read."
          },
          {
            "name": "keep",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "How close to the best candidate a name has to be to be offered alongside it, as a share of the best score. `1.0` offers only names that matched exactly as well; lower values widen the field.",
              "title": "Keep"
            },
            "description": "How close to the best candidate a name has to be to be offered alongside it, as a share of the best score. `1.0` offers only names that matched exactly as well; lower values widen the field."
          }
        ],
        "responses": {
          "200": {
            "description": "The closest names that exist, best first, or nothing at all when none of them is close enough to be worth offering.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttributeValue": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "group": {
            "$ref": "#/components/schemas/AttributeGroup"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "format": {
            "$ref": "#/components/schemas/AttributeFormat"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          },
          "derived": {
            "type": "boolean",
            "title": "Derived",
            "default": false
          },
          "prominent": {
            "type": "boolean",
            "title": "Prominent",
            "default": false
          },
          "choices": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choices"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "label",
          "group",
          "order",
          "format"
        ],
        "title": "AttributeValue",
        "description": "One field of an entity, carrying everything needed to display it without\nrecognising `key`."
      },
...
      "EntityKey": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "title": "EntityKey",
        "description": "What identifies an entity: its type, plus an id unique only within that type."
      },
...
      "EntityType": {
        "type": "string",
        "title": "EntityType",
        "description": "The types of thing this knowledge base holds: items, NPCs, shops and more. Published as an open string rather than a fixed enum, because new values get added as the data grows. The ones this build knows are listed under `x-known-values`. Treat anything else as a value you cannot render yet, not as an error.",
        "x-known-values": [
          "item",
          "npc",
          "shop",
          "quest",
          "location"
        ],
        "examples": [
          "item"
        ]
      },
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "description": "The envelope every error arrives in: one `error` object, nothing else."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hidden": {
        "properties": {
          "outcome": {
            "type": "string",
            "const": "hidden",
            "title": "Outcome",
            "default": "hidden"
          },
          "key": {
            "$ref": "#/components/schemas/EntityKey"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HiddenReason"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "Hidden",
        "description": "The entity exists in this build but is deliberately not served."
      },
...
      "JsonValue": {},
      "Link": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "title": "Slug"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "icon_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Ref"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "slug",
          "label"
        ],
        "title": "Link",
        "description": "A pointer to another entity, carrying `type`, `id`, `slug` and `label` but never\na URL."
      },
...
      "Missing": {
        "properties": {
          "outcome": {
            "type": "string",
            "const": "missing",
            "title": "Outcome",
            "default": "missing"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "reference"
        ],
        "title": "Missing",
        "description": "Nothing in this build answers to that reference."
      },
      "Moved": {
        "properties": {
          "outcome": {
            "type": "string",
            "const": "moved",
            "title": "Outcome",
            "default": "moved"
          },
          "target": {
            "$ref": "#/components/schemas/Link"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "Moved",
        "description": "The reference is retired. The entity answers at `target` now."
      },
      "PageDescriptor": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Link"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "canonical": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "array",
            "title": "Variants",
            "default": []
          },
          "infobox": {
            "items": {
              "$ref": "#/components/schemas/AttributeValue"
            },
            "type": "array",
            "title": "Infobox",
            "default": []
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "type": "array",
            "title": "Blocks",
            "default": []
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "type",
          "data_version"
        ],
        "title": "PageDescriptor",
        "description": "A whole entity page as data: identity, a line of description, the infobox, the\nsections, and a first page of each set of related entities."
      },
...
      "SearchResultPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "readOnly": true
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more",
          "next_offset"
        ],
        "title": "Page[SearchResult]"
      },
      "Present": {
        "properties": {
          "outcome": {
            "type": "string",
            "const": "found",
            "title": "Outcome",
            "default": "found"
          },
          "target": {
            "$ref": "#/components/schemas/Link"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "Present",
        "description": "The reference points at an entity that is published right now."
      },
...
      "Tooltip": {
        "properties": {
          "link": {
            "$ref": "#/components/schemas/Link"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/AttributeValue"
            },
            "type": "array",
            "title": "Attributes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "link",
          "type"
        ],
        "title": "Tooltip",
        "description": "The short form of an entity: identity, one line, and the few values the\nregistry marks as worth showing."
      },
      "TypeInfo": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "plural": {
            "type": "string",
            "title": "Plural"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/AttributeSpec"
            },
            "type": "array",
            "title": "Attributes"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipSpec"
            },
            "type": "array",
            "title": "Relationships"
          }
        },
        "type": "object",
        "required": [
          "type",
          "label",
          "plural",
          "order",
          "attributes",
          "relationships"
        ],
        "title": "TypeInfo",
        "description": "One entity type, with every attribute and relationship declared for it."
      },
...
    },
    "securitySchemes": {
      "issued_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Send the key you were issued as `Authorization: Bearer <token>`. Keys are handed out by whoever runs this deployment and do not expire; one that leaks is withdrawn by its holder's key id. A browser cannot keep one secret, so a public front end should call this through its own backend."
      }
    }
  },
...
}